Move file upload component in a separate file
