+++ client/dashboard/domains/domain-dns/import-bind-file-button.tsx
@@ -1,0 +1,53 @@
+import { useMutation } from '@tanstack/react-query';
+import { Button, FormFileUpload } from '@wordpress/components';
+import { useDispatch } from '@wordpress/data';
+import { __ } from '@wordpress/i18n';
+import { store as noticesStore } from '@wordpress/notices';
+import { domainDnsImportBindMutation } from '../../app/queries/domain-dns-records';
+import type { DnsRecord } from '../../data/domain-dns-records';
+
+interface ImportBindFileButtonProps {
+	domainName: string;
+	onRecordsImported: ( data: DnsRecord[] ) => void;
+}
+
+export default function ImportBindFileButton( {
+	domainName,
+	onRecordsImported,
+}: ImportBindFileButtonProps ) {
+	const { createErrorNotice } = useDispatch( noticesStore );
+	const importDnsBindMutation = useMutation( domainDnsImportBindMutation( domainName ) );
+
+	const handleFileChange = ( event: React.ChangeEvent< HTMLInputElement > ) => {
+		const file = event.currentTarget.files?.[ 0 ];
+		if ( ! file ) {
+			return;
+		}
+
+		const formData: [ string, File, string ][] = [ [ 'files[]', file, file.name ] ];
+		importDnsBindMutation.mutate( formData, {
+			onSuccess: ( data ) => {
+				onRecordsImported( data );
+			},
+			onError: () => {
+				createErrorNotice( __( 'Failed to import DNS records.' ), {
+					type: 'snackbar',
+				} );
+			},
+		} );
+	};
+
+	return (
+		<FormFileUpload
+			__next40pxDefaultSize
+			accept="text/plain"
+			multiple={ false }
+			onChange={ handleFileChange }
+			render={ ( { openFileDialog } ) => (
+				<Button variant="secondary" onClick={ openFileDialog }>
+					{ __( 'Import BIND file' ) }
+				</Button>
+			) }
+		/>
+	);
+}
